restructure controllers into api.go
