+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"github.com/bryx-ltd/user-service/controllers"
 	"github.com/bryx-ltd/user-service/initializers"
 	"github.com/gin-gonic/gin"
 )
@@ -14,11 +13,11 @@ func init() {
 func main() {
 	router := gin.Default()
 
-	router.POST("/users", controllers.PostUser)
-	router.PUT("/users/:id", controllers.UpdateUser)
-	router.GET("/users", controllers.GetUsers)
-	router.GET("/users/:id", controllers.GetUserByID)
-	router.DELETE("/users/:id", controllers.DeleteUser)
+	router.POST("/users", PostUser)
+	router.PUT("/users/:id", UpdateUser)
+	router.GET("/users", GetUsers)
+	router.GET("/users/:id", GetUserByID)
+	router.DELETE("/users/:id", DeleteUser)
 
 	router.Run()
 }
